working version of cubic + added test for it
