+++ tests/test_cubic.py
@@ -1,4 +1,4 @@
-from aioquic.quic.recovery import CubicCongestionControl, K_CUBIC_C, K_CUBIC_LOSS_REDUCTION_FACTOR, QuicSentPacket
+from aioquic.quic.congestion import CubicCongestionControl, K_CUBIC_C, K_CUBIC_LOSS_REDUCTION_FACTOR, QuicSentPacket
 import unittest
 
 def W_cubic(t, K, W_max):
@@ -11,17 +11,20 @@ def cube_root(x):
 class CubicTests(unittest.TestCase):
 
     def test_congestion_avoidance(self):
+        """
+        Check if the cubic implementation respects the mathematical formula defined in the rfc 9438
+        """
 
-        n = 20
+        n = 4000  # number of ms to check
 
-        W_max = 20
+        W_max = 5  # starting W_max
         K = cube_root(W_max*(1-K_CUBIC_LOSS_REDUCTION_FACTOR)/K_CUBIC_C)
         cwnd = W_max*K_CUBIC_LOSS_REDUCTION_FACTOR
 
         correct = []
 
         for i in range(n):
-            correct.append(W_cubic(i, K, W_max))
+            correct.append(W_cubic(i/1000, K, W_max))
 
         cubic = CubicCongestionControl(caller=None)
         cubic._W_max = W_max
@@ -31,14 +34,21 @@ class CubicTests(unittest.TestCase):
 
         results = []
         for i in range(n):
-            packet = QuicSentPacket(None, True, True, True, 0, 0)
-            packet.sent_bytes = 10     # won't affect results
-            cubic.on_packet_acked_timed(packet, i, 1000000)
+            cwnd = cubic.congestion_window # number of segments
+            
+
+            # simulate the reception of cwnd packets (a full window of acks)
+            for _ in range(int(cwnd)):
+                packet = QuicSentPacket(None, True, True, True, 0, 0)
+                packet.sent_bytes = 10     # won't affect results
+                rtt = 0
+                cubic.on_packet_acked_timed(packet, i/1000, rtt)
+
             results.append(cubic.congestion_window)
 
         for i in range(n):
-            self.assertAlmostEqual(correct[i], results[i])
-
+            # check if it is almost equal to the value of W_cubic
+            self.assertTrue(correct[i]*0.99 <= results[i] <= 1.01*correct[i], F"Error at {i}ms, Result={results[i]}, Expected={correct[i]}")
         
 
 if __name__ == '__main__':
